+++ test/sandbox/run.sh
@@ -2,14 +2,14 @@
 # ruby node_b/petasos_node_manager.rb
 # ruby cluster/petasos_cluster_manager.rb
 
-echo "Running in node_a"
-cd node_a
-ruby -I../../../lib ../../../bin/petasos
-echo "Running in node_b"
-cd ../node_b
-ruby -I../../../lib ../../../bin/petasos
-echo "Running in distributor"
-cd ../distributor
-echo "Copying basic config into place."
+# echo "RUNNER: Running in node_a"
+# cd node_a
+# ruby -I../../../lib ../../../bin/petasos
+# echo "RUNNER: Running in node_b"
+# cd ../node_b
+# ruby -I../../../lib ../../../bin/petasos
+echo "RUNNER: Running in distributor"
+cd distributor
+echo "RUNNER: Copying basic config into place."
 cp basic_petasos_distribution-config.yaml petasos_distribution-config.yaml
 ruby -I../../../lib ../../../bin/petasos